requirements.txt
Django==4.2
psycopg2-binary==2.9.9
azure-storage-blob==12.19.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
